+++ pkg/config/tripbot/helpers.go
@@ -24,3 +24,95 @@ func getEnv(key, fallback string) string {
 	}
 	return fallback
 }
+
+//TODO: this should load from a config file
+// IgnoredUsers are users who shouldn't be in the running for miles
+// https://twitchinsights.net/bots
+var IgnoredUsers = []string{
+	"0_applebadapple_0",
+	"angeloflight",
+	"anotherttvviewer",
+	"apricotdrupefruit",
+	"avocadobadado",
+	"commanderroot",
+	"communityshowcase",
+	"cyclemotion",
+	"electricallongboard",
+	"eubyt",
+	"feuerwehr",
+	"freddyybot",
+	"jobi_essen",
+	"konkky",
+	"logviewer",
+	"lurxx",
+	"mathgaming",
+	"mrreflector",
+	"n3td3v",
+	"nightbot",
+	"p0lizei_",
+	"slocool",
+	"streamlabs",
+	"taormina2600",
+	"teyyd",
+	"tripbot4000",
+	"unixchat",
+	"v_and_k",
+	"virgoproz",
+	"winsock",
+	"zanekyber",
+}
+
+// HelpMessages are all of the different things !help can return
+var HelpMessages = []string{
+	"!commands: List more commands you can use",
+	"!commands: List more commands you can use",
+	"!commands: List more commands you can use",
+	"!guess: Guess which state we are in",
+	"!leaderboard: See who has the most miles",
+	"!location: Get the current location",
+	"!miles: See your current miles",
+	"!report: Report a stream issue (frozen, no audio, etc)",
+	"!state: Get the state we are currently in",
+	"!sunset: Get time until sunset (on the day of filming)",
+	"!survey: Fill out a survey and help the stream",
+	"!timewarp: Magically warp to a new moment in time",
+	// "!song: Get the current music",
+}
+
+var GoogleMapsStyle = []string{
+	"element:geometry|color:0x242f3e",
+	"element:labels.text.fill|color:0x746855",
+	"element:labels.text.stroke|color:0x242f3e",
+	"feature:administrative.locality|element:labels.text.fill|color:0xd59563",
+	"feature:poi|element:labels.text.fill|color:0xd59563",
+	"feature:poi.park|element:geometry|color:0x263c3f",
+	"feature:poi.park|element:labels.text.fill|color:0x6b9a76",
+	"feature:road|element:geometry|color:0x38414e",
+	"feature:road|element:geometry.stroke|color:0x212a37",
+	"feature:road|element:labels.text.fill|color:0x9ca5b3",
+	"feature:road.highway|element:geometry|color:0x746855",
+	"feature:road.highway|element:geometry.stroke|color:0x1f2835",
+	"feature:road.highway|element:labels.text.fill|color:0xf3d19c",
+	"feature:transit|element:geometry|color:0x2f3948",
+	"feature:transit.station|element:labels.text.fill|color:0xd59563",
+	"feature:water|element:geometry|color:0x17263c",
+	"feature:water|element:labels.text.fill|color:0x515c6d",
+	"feature:water|element:labels.text.stroke|color:0x17263c&size=480x360",
+}
+
+// these are different timestamps we have screenshots prepared for
+// the "000" corresponds to 0m0s, "130" corresponds to 1m30s
+var TimestampsToTry = []string{
+	"000",
+	"015",
+	"030",
+	"045",
+	"100",
+	"115",
+	"130",
+	"145",
+	"200",
+	"215",
+	"230",
+	"245",
+}
